+++ core/src/main/java/io/questdb/cairo/vm/MemoryCMRImpl.java
@@ -40,7 +40,7 @@ import static io.questdb.cairo.vm.Vm.PARANOIA_MODE;
 //contiguous mapped readable 
 public class MemoryCMRImpl extends AbstractMemoryCR implements MemoryCMR {
     private static final Log LOG = LogFactory.getLog(MemoryCMRImpl.class);
-    protected int fd = -1;
+    protected long fd = -1;
     protected int memoryTag = MemoryTag.MMAP_DEFAULT;
     private int madviseOpts = -1;
 
@@ -76,8 +76,8 @@ public class MemoryCMRImpl extends AbstractMemoryCR implements MemoryCMR {
     }
 
     @Override
-    public int detachFdClose() {
-        int fd = this.fd;
+    public long detachFdClose() {
+        long fd = this.fd;
         this.fd = -1;
         close();
         return fd;
@@ -91,7 +91,7 @@ public class MemoryCMRImpl extends AbstractMemoryCR implements MemoryCMR {
     }
 
     @Override
-    public int getFd() {
+    public long getFd() {
         return fd;
     }
