+++ core/src/main/java/io/questdb/std/FilesFacadeImpl.java
@@ -40,7 +40,7 @@ public class FilesFacadeImpl implements FilesFacade {
     private long mapPageSize = 0;
 
     @Override
-    public boolean allocate(int fd, long size) {
+    public boolean allocate(long fd, long size) {
         return Files.allocate(fd, size);
     }
 
@@ -65,17 +65,17 @@ public class FilesFacadeImpl implements FilesFacade {
     }
 
     @Override
-    public long append(int fd, long buf, int len) {
+    public long append(long fd, long buf, int len) {
         return Files.append(fd, buf, len);
     }
 
     @Override
-    public boolean close(int fd) {
+    public boolean close(long fd) {
         return Files.close(fd) == 0;
     }
 
     @Override
-    public boolean closeRemove(int fd, LPSZ path) {
+    public boolean closeRemove(long fd, LPSZ path) {
         // On Windows we cannot remove file that is open, close it first
         if (isRestrictedFileSystem() && fd > -1) {
             Files.close(fd);
@@ -96,12 +96,12 @@ public class FilesFacadeImpl implements FilesFacade {
     }
 
     @Override
-    public long copyData(int srcFd, int destFd, long offsetSrc, long length) {
+    public long copyData(long srcFd, long destFd, long offsetSrc, long length) {
         return Files.copyData(srcFd, destFd, offsetSrc, length);
     }
 
     @Override
-    public long copyData(int srcFd, int destFd, long offsetSrc, long destOffset, long length) {
+    public long copyData(long srcFd, long destFd, long offsetSrc, long destOffset, long length) {
         return Files.copyDataToOffset(srcFd, destFd, offsetSrc, destOffset, length);
     }
 
@@ -121,12 +121,12 @@ public class FilesFacadeImpl implements FilesFacade {
     }
 
     @Override
-    public boolean exists(int fd) {
+    public boolean exists(long fd) {
         return Files.exists(fd);
     }
 
     @Override
-    public void fadvise(int fd, long offset, long len, int advise) {
+    public void fadvise(long fd, long offset, long len, int advise) {
         if (advise > -1) {
             Files.fadvise(fd, offset, len, advise);
         }
@@ -169,7 +169,7 @@ public class FilesFacadeImpl implements FilesFacade {
     }
 
     @Override
-    public void fsync(int fd) {
+    public void fsync(long fd) {
         int res = Files.fsync(fd);
         if (res == 0) {
             return;
@@ -178,7 +178,7 @@ public class FilesFacadeImpl implements FilesFacade {
     }
 
     @Override
-    public void fsyncAndClose(int fd) {
+    public void fsyncAndClose(long fd) {
         int res = Files.fsync(fd);
         if (res == 0) {
             close(fd);
@@ -291,7 +291,7 @@ public class FilesFacadeImpl implements FilesFacade {
     }
 
     @Override
-    public long length(int fd) {
+    public long length(long fd) {
         long r = Files.length(fd);
         if (r < 0) {
             throw CairoException.critical(Os.errno()).put("Checking file size failed");
@@ -305,7 +305,7 @@ public class FilesFacadeImpl implements FilesFacade {
     }
 
     @Override
-    public int lock(int fd) {
+    public int lock(long fd) {
         return Files.lock(fd);
     }
 
@@ -327,12 +327,12 @@ public class FilesFacadeImpl implements FilesFacade {
     }
 
     @Override
-    public long mmap(int fd, long len, long offset, int flags, int memoryTag) {
+    public long mmap(long fd, long len, long offset, int flags, int memoryTag) {
         return Files.mmap(fd, len, offset, flags, memoryTag);
     }
 
     @Override
-    public long mremap(int fd, long addr, long previousSize, long newSize, long offset, int mode, int memoryTag) {
+    public long mremap(long fd, long addr, long previousSize, long newSize, long offset, int mode, int memoryTag) {
         return Files.mremap(fd, addr, previousSize, newSize, offset, mode, memoryTag);
     }
 
@@ -351,12 +351,12 @@ public class FilesFacadeImpl implements FilesFacade {
     }
 
     @Override
-    public int openAppend(LPSZ name) {
+    public long openAppend(LPSZ name) {
         return Files.openAppend(name);
     }
 
     @Override
-    public int openCleanRW(LPSZ name, long size) {
+    public long openCleanRW(LPSZ name, long size) {
         // Open files and if file exists, try exclusively lock it
         // If exclusive lock worked the file will be cleaned and allocated to the given size
         // Shared lock will be left on the file which will be removed when file descriptor is closed
@@ -365,22 +365,22 @@ public class FilesFacadeImpl implements FilesFacade {
     }
 
     @Override
-    public int openRO(LPSZ name) {
+    public long openRO(LPSZ name) {
         return Files.openRO(name);
     }
 
     @Override
-    public int openRW(LPSZ name, long opts) {
+    public long openRW(LPSZ name, long opts) {
         return Files.openRW(name, opts);
     }
 
     @Override
-    public long read(int fd, long buf, long len, long offset) {
+    public long read(long fd, long buf, long len, long offset) {
         return Files.read(fd, buf, len, offset);
     }
 
     @Override
-    public long readIntAsUnsignedLong(int fd, long offset) {
+    public long readIntAsUnsignedLong(long fd, long offset) {
         return Files.readIntAsUnsignedLong(fd, offset);
     }
 
@@ -390,17 +390,17 @@ public class FilesFacadeImpl implements FilesFacade {
     }
 
     @Override
-    public byte readNonNegativeByte(int fd, long offset) {
+    public byte readNonNegativeByte(long fd, long offset) {
         return Files.readNonNegativeByte(fd, offset);
     }
 
     @Override
-    public int readNonNegativeInt(int fd, long offset) {
+    public int readNonNegativeInt(long fd, long offset) {
         return Files.readNonNegativeInt(fd, offset);
     }
 
     @Override
-    public long readNonNegativeLong(int fd, long offset) {
+    public long readNonNegativeLong(long fd, long offset) {
         return Files.readNonNegativeLong(fd, offset);
     }
 
@@ -457,7 +457,7 @@ public class FilesFacadeImpl implements FilesFacade {
     }
 
     @Override
-    public boolean truncate(int fd, long size) {
+    public boolean truncate(long fd, long size) {
         return Files.truncate(fd, size);
     }
 
@@ -505,7 +505,7 @@ public class FilesFacadeImpl implements FilesFacade {
     }
 
     @Override
-    public long write(int fd, long address, long len, long offset) {
+    public long write(long fd, long address, long len, long offset) {
         return Files.write(fd, address, len, offset);
     }
