+++ core/src/test/java/io/questdb/test/cutlass/line/tcp/LineTcpConnectionContextTest.java
@@ -450,7 +450,7 @@ public class LineTcpConnectionContextTest extends BaseLineTcpContextTest {
         runInContext(
                 new TestFilesFacadeImpl() {
                     @Override
-                    public int openRW(LPSZ name, long opts) {
+                    public long openRW(LPSZ name, long opts) {
                         if (Utf8s.endsWithAscii(name, "broken.d.1")) {
                             return -1;
                         }
@@ -490,7 +490,7 @@ public class LineTcpConnectionContextTest extends BaseLineTcpContextTest {
         runInContext(
                 new TestFilesFacadeImpl() {
                     @Override
-                    public int openRW(LPSZ name, long opts) {
+                    public long openRW(LPSZ name, long opts) {
                         if (Utf8s.endsWithAscii(name, "1970-01-01.1" + Files.SEPARATOR + "temperature.d")) {
                             return -1;
                         }
@@ -528,7 +528,7 @@ public class LineTcpConnectionContextTest extends BaseLineTcpContextTest {
         runInContext(
                 new TestFilesFacadeImpl() {
                     @Override
-                    public int openRW(LPSZ name, long opts) {
+                    public long openRW(LPSZ name, long opts) {
                         if (Utf8s.endsWithAscii(name, "broken.d")) {
                             return -1;
                         }
